BugFix: Removing Mandatory Attributes on Federation Creation
Major change is to make the role attribute not mandatory anymore, as the OIDC10 Protocol doesn't require/allow this because it is an OIDC relying party by definition.

Other attributes (protocol, configuration) are also made optional in ansible to allow all kind of flexibility with these parameters. They are still required by ibmsecurity ! (current state)